+++ packages/unifi-mcp-shared/src/unifi_mcp_shared/formatting.py
@@ -1,17 +1,2 @@
-"""Standardized tool response formatting."""
-
-from typing import Any
-
-
-def success_response(data: Any = None, **kwargs) -> dict[str, Any]:
-    result = {"success": True}
-    if data is not None:
-        result["data"] = data
-    result.update(kwargs)
-    return result
-
-
-def error_response(error: str, **kwargs) -> dict[str, Any]:
-    result = {"success": False, "error": error}
-    result.update(kwargs)
-    return result
+"""Compatibility shim. Real implementation lives in unifi_core.formatting."""
+from unifi_core.formatting import *  # noqa: F401,F403
